Add Twitter's Bootstrap 3.0.0 CSS and Javascript files, under Apache License 2.0
These files are exactly as they appear the upstream release 3.0.0 of
Bootstrap, which Twitter released under the Apache License 2.0.  To extract
these files, I did the following:

I downloaded the following file:
  https://github.com/twbs/bootstrap/archive/v3.0.0.zip

with sha256sum of:
$ sha256sum v3.0.0.zip
   2d54f345f4abc6bf65ea648c323e9bae577e6febf755650e62555f2d7a222e17  v3.0.0.zip

And extracted from it these two files:
   bootstrap-3.0.0/dist/css/bootstrap.css
   bootstrap-3.0.0/dist/js/bootstrap.js
which are licensed under the Apache License 2.0.

and placed them into:
  rhodecode/public/css/bootstrap.css
  rhodecode/public/js/bootstrap.js
respectively.