pull requests commit overview: respect newlines and indentation (Issue #79)
Before this commit, commit messages in the pull request commit overview
would be handled like any other text in HTML, with newlines and multiple
whitespace being irrelevant.

Remove an overrule in the CSS so that the 'white-space' property remains at
'pre-wrap', causing newlines and indentation in commit messages are properly
respected in this overview. As a positive effect, only the first line of the
commit message is shown by default, just like in standard version control
commands like 'hg log'.