pullrequest overview: display in table to improve readability
Instead of showing all info in the pullrequest overview right after each
other on the same line, use a table to improve readability.

Additionally, add following information:
- author
- destination repository

The latter is always the same when viewing the list of pullrequests of a
certain repository, but this change already anticipates re-use of this
block in the 'my pullrequests' overview.

The table is still static, non-sortable, since YUI is deprecated and no new
alternative has been decided upon yet. This improvement can be done later.