summary: repeat repo type in front of clone URL
Repeatedly, users report cloning problems of the Kallithea sources on IRC,
caused by them assuming that it is a Git repository.

Try to solve this problem by repeating the repository tag, indicating 'hg'
or 'git' in front of the clone URL.

Additionally, decrease the font size of the URL a notch, and improve
alignment with the repotag by removing top/bottom padding.

To avoid duplication of the repotag logic, introduce a mako def for it.