tests: hide database setup queries with pytest
Unlike nosetest, pytest would show the database setup queries at the
beginning of the test run, which don't bring added value.
Hide them by disabling logging during this time.

Note that this does not provide an easy method of enabling the logging on
demand (what could be done with the -s switch in nosetest).