comments: previous/next links should only be to first comment for a given line
Before, each comment had its own link to previous/next comment - also if it was
right before/after in the same thread. Comments right before/after in the same
thread are however easy to navigate to with scrolling or arrow keys or
pageup/down. Navigating to the next thread could take a lot of clicks to cycle
through every single comment ever made.

Instead, just show the link at the top of each thread and let the links
navigate between threads.

This also make long threads use less vertical space.

One downside of this change is that it for long threads takes a bit more work
to get from the last comment in a thread to the next thread.