docs: update example output and example server configs
kallithea-api example output was not up-to-date, and the text was a
little vague on whether you specify a hostname or a URL.

The Nginx example config has been updated to assume a Kallithea backend
server on localhost:5000, like the Apache example.

The redundant ServerAlias option was removed from the Apache example.