db: make sure all (non-primary) columns have nullable set explicitly
The default of nullable=True is rarely good for us so nullable should always be
specified unless there is a reason to allow nullable ... and if the default is
fine, xplicit is better than implicit.

The declared nulliness of some fields are changed where it seems like code
already enforced it.

Some fields are marked as FIXME when they need (trivial?) data conversion to
convert NULLs to default values.