test_admin_notifications: fix index test dependency
The index test only worked because another test had prepared some database
state it relied on.

More specifically, the index test creates a notification
as a newly created user, but that user had not been committed to the
database yet. When running the index test standalone, this causes the error:

    IntegrityError: (IntegrityError) NOT NULL constraint failed:
    notifications.created_by u'INSERT INTO notifications (subject, body,
    created_by, created_on, type) VALUES (?, ?, ?, ?, ?)'
    (u'test_notification_1', u'notification_1', None, '2016-03-11
    21:15:19.066402', u'message')

By committing the database after the user is created, the problem is solved.