pull requests: prevent adding DEFAULT user as reviewer
Add a helper method to resolve reviewers, with an added check to prevent
the adding of the DEFAULT user as reviewer.

The __add_reviewers method, although internal, had a troubling interface
where the method was responsible for resolving reviewers, but the caller
was responsible for resolving mention_recipients. The method is changed
to always take two sets of database User objects, leaving the resolution
to the caller in both cases.