test.ini: align some logging-related settings with development.ini
When manually using test.ini (that is, not via the test suite) the coloring
of log output like development.ini provides can be very useful.
Note that this color output is not taken along by pytest-capturelog,
regardless of this patch.

Some extra-verbose logging in test.ini is on the other hand not needed and
could be enabled when needed by a specific developer.

Note: the 'level=DEBUG' setting for handler_console_sql is not taken along
to test.ini: it causes duplicate sqlalchemy debug logs, one through
handler_console_sql and another through another path.

If future experience points out that some of these changes are actually
worse than they were, adjustments can still be made.