tests: admin_users: make sure repo group is deleted
test_delete_repo_group_err creates then deletes a repository group. However,
if the delete fails the repository group remains. This later causes problems
in the model tests.

Introduce a pytest yield fixture to handle the creation _and_ deletion of
the repository group (suggested by S?ren L?vborg).
The creation of the user needs to happen _before_ that of the user group,
and we cannot share data between two pytest fixtures, so the user is created
in the fixture as well.