controllers: remove redundant Repository database lookups
BaseRepoController already provides the repository object in c.db_repo
(may be None) and BaseController already provides the repository name
in c.repo_name (as given in the URL).

(Arguably, that's a bad design, and we should revisit that decision in
the future. For now, the code just performs slightly better.)