comments: for PR comments, link directly to comment on PR page
Comments with approval changes made in the context of PRs also show up
on the individual PR changesets, and currently link to the (top of the)
PR page. If the user wants to see the surrounding comments, it can be
difficult to find the comment.

This change makes the link go to the comment anchor on the PR page, so
the user immediately sees the relevant comment thread. The PR metadata
is still readily available (by scrolling up or pressing "Home").